[EDITED : AC_SkeletonFollower.cpp + .h]
- changed movement behaviour of skeleton :

circle is drawn on both players,
nearest point on circle to skeleton, on both players is found,
center if those 2 points is set as a new spline point,
a spline is drawn from the skeleton to this new spline point,
skeleton follows spline.

navmesh :
if there is an object that is considered a "collision" on the skeletons channel is on this splines path,
retrace backwards and nudge the point of the spline along backwards from its original point, to the skeleton,
once the spline point is outside of this collisions border, set spline point, draw spline, and make skeleton follow along.